refactor: rename files in lib/interfaces for temporary since its not saving the name (eg. cat.ts to Cat.ts) file to the commit
