+++ lib/interfaces/catDetails_temp.ts
