+++ lib/interfaces/catStats_temp.ts
