+++ lib/interfaces/cat_temp.ts
