+++ lib/interfaces/perCatNotificationPref_temp.ts
